refactor(metrics): extract scorecard module
Move Benchmark, ScorecardRoute, load_benchmarks, scorecard_routes, and all
their tests from mod.rs into a new scorecard.rs; mod.rs is now a pure
10-line re-export layer.
